refactor category resource
