+++ tests/unit/test_category.py
@@ -4,7 +4,7 @@ import responses
 
 from tamr_unify_client import Client
 from tamr_unify_client.auth import UsernamePasswordAuth
-from tamr_unify_client.models.taxonomy.category import Category
+from tamr_unify_client.models.category.resource import Category
 
 
 class TestCategory(TestCase):
