+++ tests/unit/test_taxonomy.py
@@ -4,8 +4,8 @@ import responses
 
 from tamr_unify_client import Client
 from tamr_unify_client.auth import UsernamePasswordAuth
-from tamr_unify_client.models.taxonomy.category import Category
-from tamr_unify_client.models.taxonomy.category_collection import CategoryCollection
+from tamr_unify_client.models.category.resource import Category
+from tamr_unify_client.models.category.collection import CategoryCollection
 from tamr_unify_client.models.taxonomy.resource import Taxonomy
